fix bottom of page links
in MKDocs style, when linking to an .md, we don't have to supply a
(fully) qualified URL, MKDocs itself can resolve the link properly.

Note that we're still missing the link to

```
[Advanced Messaging](advanced)
```

That page is missing.

Note that this PR also introduces
[`.gitattributes`](https://www.git-scm.com/docs/gitattributes), like we
did in https://github.com/riker-rs/riker/pull/126 because my editor
inadvarteldy introduced LF here, and trying to fix it made it worse.